Merging David's f706630c commit.
The start URL in rhoconfig assumes that the start URL of *all* the tabs are what is specified in rhoconfig.txt. However the @tabs array (a.k.a. TabData) is what really determines what the start URL is on a tabbed application. Changed the MainView interface to provide start URL and whether or not we are on the start tab. The MainView uses the original implementation, while the TabbedMainView compares against TabData instead. This solves a problem where the back button would create an empty page if pressed right after opening the app on all the tabs.

Conflicts:

	platform/android/Rhodes/src/com/rhomobile/rhodes/SplashScreen.java