Delegate class support
Read the delegate class from a file and use it to forward any commands that we
are not aware of. That way we can use it to forward commands. The class should be
defined on rhoconfig's iphone_splash_delegate_class, and it can live on any
extension that is linked directly to the executable.